requirements.txt
# 핵심 라이브러리
requests>=2.28.0
msal>=1.20.0
click>=8.1.3
rich>=12.6.0
python-dotenv>=0.21.0

# 데이터 처리
beautifulsoup4>=4.11.1  # HTML 처리
html2text>=2020.1.16    # HTML -> 텍스트 변환

# 테스트
pytest>=7.2.0

# 타입 힌트 및 정적 타입 검사
mypy>=0.991
types-requests
types-python-dateutil

# 개발 도구
black>=22.12.0  # 코드 포맷팅
flake8>=6.0.0   # 코드 린팅
